+++ R/MotrpacRatTraining6mo-package.R
@@ -5,7 +5,10 @@
 #' @import MotrpacRatTraining6moData
 #' @importFrom data.table as.data.table data.table rbindlist setnames := copy melt setkey dcast
 #' @importFrom magrittr %>%
-#' @importFrom ggplot2 guides scale_size theme annotate guide_legend aes element_rect ggplot geom_bar theme_classic facet_wrap scale_fill_manual coord_flip labs scale_x_discrete element_blank geom_line
+#' @importFrom ggplot2 guides scale_size theme annotate guide_legend aes element_rect 
+#'   ggplot geom_bar theme_classic facet_wrap scale_fill_manual coord_flip labs 
+#'   scale_x_discrete element_blank geom_line geom_point geom_errorbar scale_colour_manual
+#'   element_text margin unit
 #' @importFrom stats na.omit prcomp sd median poly phyper cor p.adjust coefficients
 #' @importFrom utils download.file data globalVariables combn
 #' @importFrom graphics hist legend