+++ NEWS.md
@@ -1,6 +1,7 @@
 # MotrpacRatTraining6mo 1.3.0 (2023-01-?)
 
-* 
+* Add `df_to_numeric()` to easily format data frames. 
+* Add `plot_feature_normalized_data()` to plot normalized sample-level data for a single feature. 
 
 # MotrpacRatTraining6mo 1.2.0 (2022-12-27)